Move tests into existing test_s3.py and example_s3.py
Address PR review feedback:
- Move unit tests into test_s3.py (1:1 matching with s3.py)
- Add read_object task to existing example_s3.py system test
- Update docs to reference example_s3.py
- Delete separate test files
