+++ providers/amazon/tests/unit/amazon/aws/operators/test_s3.py
@@ -46,6 +46,7 @@ from airflow.providers.amazon.aws.operators.s3 import (
     S3ListOperator,
     S3ListPrefixesOperator,
     S3PutBucketTaggingOperator,
+    S3ReadObjectOperator,
 )
 from airflow.providers.common.compat.openlineage.facet import (
     Dataset,
@@ -1083,3 +1084,66 @@ class TestS3CreateObjectOperator:
     def test_template_fields(self):
         operator = S3CreateObjectOperator(task_id="test", s3_bucket="bucket", s3_key="key", data="test")
         validate_template_fields(operator)
+
+
+class TestS3ReadObjectOperator:
+    def setup_method(self):
+        self.operator = S3ReadObjectOperator(
+            task_id="test-s3-read-object",
+            s3_bucket=BUCKET_NAME,
+            s3_key=S3_KEY,
+        )
+
+    @mock_aws
+    def test_execute_reads_object(self):
+        conn = boto3.client("s3", region_name="us-east-1")
+        conn.create_bucket(Bucket=BUCKET_NAME)
+        conn.put_object(Bucket=BUCKET_NAME, Key=S3_KEY, Body=b"hello world from s3")
+
+        result = self.operator.execute({})
+        assert result == "hello world from s3"
+
+    @mock_aws
+    @mock.patch.object(S3Hook, "read_key")
+    def test_execute_calls_read_key(self, mock_read_key):
+        mock_read_key.return_value = "content"
+        result = self.operator.execute({})
+        mock_read_key.assert_called_once_with(key=S3_KEY, bucket_name=BUCKET_NAME)
+        assert result == "content"
+
+    @mock_aws
+    def test_execute_with_s3_url(self):
+        conn = boto3.client("s3", region_name="us-east-1")
+        conn.create_bucket(Bucket=BUCKET_NAME)
+        conn.put_object(Bucket=BUCKET_NAME, Key=S3_KEY, Body=b"url-content")
+
+        op = S3ReadObjectOperator(
+            task_id="test-s3-url",
+            s3_key=f"s3://{BUCKET_NAME}/{S3_KEY}",
+        )
+        result = op.execute({})
+        assert result == "url-content"
+
+    @mock_aws
+    def test_execute_empty_object(self):
+        conn = boto3.client("s3", region_name="us-east-1")
+        conn.create_bucket(Bucket=BUCKET_NAME)
+        conn.put_object(Bucket=BUCKET_NAME, Key=S3_KEY, Body=b"")
+
+        result = self.operator.execute({})
+        assert result == ""
+
+    @pytest.mark.parametrize(("bucket", "key"), (("bucket", "file.txt"), (None, "s3://bucket/file.txt")))
+    def test_get_openlineage_facets_on_start(self, bucket, key):
+        expected_input = Dataset(
+            namespace="s3://bucket",
+            name="file.txt",
+        )
+        op = S3ReadObjectOperator(task_id="test", s3_bucket=bucket, s3_key=key)
+        lineage = op.get_openlineage_facets_on_start()
+        assert len(lineage.outputs) == 0
+        assert len(lineage.inputs) == 1
+        assert lineage.inputs[0] == expected_input
+
+    def test_template_fields(self):
+        validate_template_fields(self.operator)
